fix: 체크리스트 가져오는 쿼리 결과를 객체에 연결 시 발생하는 오류를 해결한다
 - Checklist 클래스에 잘못 정의된 필드 타입을 DB에 맞게 고친다
 - snake case로 정의된 DB 컬럼명을 camel case로 제대로 바꾸지 못하는
   문제를 Mybatis 설정(map-underscore-to-camel-case)으로 해결한다
 - 추가한 테스트용 API uri("/assigned")를 설정해 기존 GET API와 구분한다
 - 기존 코드에 변경사항을 반영한다

+++ src/main/java/org/swyp/weddy/domain/checklist/service/ChecklistService.java
@@ -4,7 +4,7 @@ import org.swyp.weddy.domain.checklist.service.dto.ChecklistDto;
 import org.swyp.weddy.domain.checklist.web.response.ChecklistResponse;
 
 public interface ChecklistService {
-    int assignChecklist(ChecklistDto dto);
+    Long assignChecklist(ChecklistDto dto);
 
     boolean hasChecklist(ChecklistDto dto);
